+++ src/index.tsx
@@ -1,7 +1,12 @@
 import * as React from 'react';
 import * as ReactDOM from 'react-dom';
-const root = document.getElementById('root');
+import Counter from '@/components/Counter';
+import Todos from '@/components/Todos';
 
-const props = { className: 'title' };
-const element = React.createElement('div', props, 'hello');
-ReactDOM.render(element, root);
+const root = document.getElementById('root');
+ReactDOM.render((
+  <>
+    <Counter number={100}/>
+    <Todos title="待办事项"/>
+  </>
+), root);