+++ webpack.config.js
@@ -1,33 +1,36 @@
 const webpack = require('webpack');
 const HtmlWebpackPlugin = require('html-webpack-plugin');
 const path = require('path');
+const { mode } = require('webpack-nano/argv');
+const { WebpackPluginServe } = require('webpack-plugin-serve');
 
 module.exports = {
-  mode: "development",
+  watch: mode === 'development',
+  mode: 'development',
   devtool: false,
-  entry: "./src/index.tsx",
+  entry: './src/index.tsx',
   output: {
-    filename: "[name].[hash].js",
-    path: path.join(__dirname, "dist")
+    filename: '[name].[hash].js',
+    path: path.join(__dirname, 'dist')
   },
   devServer: {
     hot: true,
-    contentBase: path.join(__dirname, "dist"),
+    contentBase: path.join(__dirname, 'dist'),
     historyApiFallback: {
       index: './index.html'
     }
   },
   resolve: {
-    extensions: [".ts", ".tsx", ".js", ".json"],
+    extensions: ['.ts', '.tsx', '.js', '.json'],
     alias: {
-      "@": path.resolve("src"),
+      '@': path.resolve('src'),
     }
   },
   module: {
     rules: [
       {
         test: /\.tsx?$/,
-        loader: "ts-loader"
+        loader: 'ts-loader'
       }
     ]
   },
@@ -35,6 +38,12 @@ module.exports = {
     new HtmlWebpackPlugin({
       template: './src/index.html'
     }),
-    new webpack.HotModuleReplacementPlugin()
+    // new webpack.HotModuleReplacementPlugin(),
+    new WebpackPluginServe({
+      port: process.env.PORT || 8080,
+      static: './dist',
+      liveReload: true,
+      waitForBuild: true,
+    }),
   ]
-}
+};