+++ CHANGELOG.md
@@ -1,8 +1,12 @@
-# 1.0.0 (2020-12-20)
+# 1.0.0 (2020-12-21)
 
 
 ### Features
 
+* add eslint & jest ([2c32ada](https://github.com/weisuoke/zf-ts-react/commit/2c32adaaddd481de1adbbf7bf552767e9ef2517e))
+* add travis ([f51376a](https://github.com/weisuoke/zf-ts-react/commit/f51376a45a35eef8ffc01a77c14425711941a2e7))
+* config ts and webpack ([a04a5f9](https://github.com/weisuoke/zf-ts-react/commit/a04a5f9003dd95e48b009b9486a40f7d62ff8f46))
+* config ts and webpack ([0890447](https://github.com/weisuoke/zf-ts-react/commit/0890447b6890d5a5e6cc36ae66eeb3f3486ec3d8))
 * project init ([454b9f9](https://github.com/weisuoke/zf-ts-react/commit/454b9f9b583197eafc7c6a0915f25885a4a60132))